update test watcher usage in base class
