+++ best-practice/src/test/java/com/saucedemo/MobileTestsBase.java
@@ -1,0 +1,9 @@
+package com.saucedemo;
+
+import io.appium.java_client.AppiumDriver;
+
+public class MobileTestsBase extends TestBase {
+    public AppiumDriver getDriver() {
+        return (AppiumDriver) driver;
+    }
+}
