--- best-practice/src/test/java/com/saucedemo/WebTestsBase.java
@@ -1,11 +1,0 @@
-package com.saucedemo;
-
-import org.openqa.selenium.remote.RemoteWebDriver;
-
-public class WebTestsBase extends TestBase {
-    public RemoteWebDriver driver;
-
-    public RemoteWebDriver getDriver() {
-        return driver;
-    }
-}
